Update gedcom.py
Hi Team.

I added the family description part based on the existing code.
The codes worked. I was able to run it using my the GED file if I input the file path directly in the codes. But for some reason it is not taking in file if it passed through the terminal.